functions/db-query-index/requirements.txt
google-cloud-storage==1.9.0
google-cloud-pubsub==0.30.1
py2neo==4.0.0
neo4j-driver==1.6.2